약관 로직 구현중
